refactor: co-locate pedigree-layout with FamilyPedigree stage
Move lib/pedigree-layout into lib/interviewer/Interfaces/FamilyPedigree
since the FamilyPedigree stage is its only consumer.
