+++ lib/interviewer/Interfaces/FamilyPedigree/pedigree-layout/__tests__/utils.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { type ParentConnection } from '~/lib/pedigree-layout/types';
+import { type ParentConnection } from '~/lib/interviewer/Interfaces/FamilyPedigree/pedigree-layout/types';
 import {
   ancestor,
   chaseup,
@@ -11,7 +11,7 @@ import {
   rank,
   tableCounts,
   which,
-} from '~/lib/pedigree-layout/utils';
+} from '~/lib/interviewer/Interfaces/FamilyPedigree/pedigree-layout/utils';
 
 describe('createMatrix', () => {
   it('creates a 2D array of correct dimensions', () => {
