+++ lib/interviewer/Interfaces/FamilyPedigree/pedigree-layout/components/FamilyPedigreePlaceholder.tsx
