+++ lib/interviewer/Interfaces/FamilyPedigree/pedigree-layout/utils/getDisplayLabel.ts
